Swap forward-delete and equal/plus
It's significantly more common for me to need plus/equals than forward
delete, so putting it in an easier place will help a lot. This also
mirrors the right side where minus is in the top right, so this kind of
mirrors the opposite on the left side, for symmetry.